requirements.txt
minecraft-launcher-lib>=8.0
requests>=2.32.0
Pillow>=10.0.0

